+++ src/config/index.ts
@@ -1,14 +1,22 @@
-import * as child from 'child_process';
+/*
+sfdx-hardis is managed in 3 layers, wit hthe following priority
+- project, stored in /config
+- branches, stored in /config/branches
+- user, stored in /config/users
+
+getConfig(layer) returns:
+- project + branches + user if layer is user
+- project + branches if layer is branch
+- project if layer is project
+*/
+
 import { cosmiconfig } from 'cosmiconfig';
 import * as fs from 'fs-extra';
 import * as yaml from 'js-yaml';
 import * as os from 'os';
-import simpleGit, {SimpleGit} from 'simple-git';
-let git: SimpleGit = null ;
-const isGit = child.exec('git rev-parse --is-inside-work-tree 2>/dev/null', {encoding: 'utf8'});
-if (isGit) {
-    git = simpleGit();
-}
+import * as path from 'path';
+import * as prompts from 'prompts';
+import { getCurrentGitBranch, isGit } from '../common/utils';
 
 const moduleName = 'sfdx-hardis';
 const projectConfigFiles = [
@@ -20,36 +28,35 @@ const projectConfigFiles = [
 ];
 const username = os.userInfo().username;
 const userConfigFiles = [
-    `config/.${moduleName}.${username}.yaml`,
-    `config/.${moduleName}.${username}.yml`
+    `config/user/.${moduleName}.${username}.yaml`,
+    `config/user/.${moduleName}.${username}.yml`
 ];
 
 async function getBranchConfigFiles() {
     if (!isGit) {
         return [];
     }
-    const gitBranch = (await git.branchLocal()).current;
-    const gitBranchFormatted = gitBranch.replace('/', '-');
+    const gitBranchFormatted = await getCurrentGitBranch({formatted: true});
     const branchConfigFiles = [
-        `config/.${moduleName}.${gitBranchFormatted}.yaml`,
-        `config/.${moduleName}.${gitBranchFormatted}.yml`
+        `config/branches/.${moduleName}.${gitBranchFormatted}.yaml`,
+        `config/branches/.${moduleName}.${gitBranchFormatted}.yml`
     ];
     return branchConfigFiles;
 }
 
-export const getConfig = async (layer: string = 'project'): Promise<any> => {
+export const getConfig = async (layer: string = 'user'): Promise<any> => {
     const defaultConfig = await loadFromConfigFile(projectConfigFiles);
     if (layer === 'project') {
         return defaultConfig;
     }
-    let userConfig = await loadFromConfigFile(userConfigFiles);
-    userConfig = Object.assign(defaultConfig, userConfig);
-    if (layer === 'user') {
-        return userConfig;
-    }
     let branchConfig = await loadFromConfigFile(await getBranchConfigFiles());
     branchConfig = Object.assign(defaultConfig, branchConfig);
-    return branchConfig;
+    if (layer === 'branch') {
+        return branchConfig;
+    }
+    let userConfig = await loadFromConfigFile(userConfigFiles);
+    userConfig = Object.assign(defaultConfig, userConfig);
+    return userConfig;
 };
 
 // Set data in configuration file
@@ -73,12 +80,61 @@ async function loadFromConfigFile(searchPlaces: string[]): Promise<any> {
 
 // Update configuration file
 async function setInConfigFile(searchPlaces: string[], propValues: any) {
-    const configExplorer = await cosmiconfig(moduleName, { searchPlaces }).search();
+    const explorer = cosmiconfig(moduleName, { searchPlaces });
+    const configExplorer = await explorer.search();
     const configFile = (configExplorer != null) ? configExplorer.filepath : searchPlaces.slice(-1)[0];
     let doc = {};
     if (fs.existsSync(configFile)) {
-        doc = yaml.load(fs.readFileSync(configFile, 'utf8'));
+        doc = yaml.load(fs.readFileSync(configFile, 'utf-8'));
     }
     doc = Object.assign(doc, propValues);
+    await fs.ensureDir(path.dirname(configFile));
     await fs.writeFile(configFile, yaml.dump(doc));
+    explorer.clearCaches();
 }
+
+// Check configuration of project so it works with sfdx-hardis
+export  const checkConfig = async (options: any) => {
+    // Skip hooks from other commands than hardis:scratch commands
+    const commandId = options?.Command?.id || options?.id ||  '';
+    if (!commandId.startsWith('hardis')) {
+        return;
+    }
+
+    let devHubAliasOk = false;
+    // Check projectName is set. If not, request user to input it
+    if (
+        options.Command &&
+        (options.Command.requiresProject === true || options.Command.supportsDevhubUsername === true)
+    ) {
+        const configProject = await getConfig('project');
+        let projectName = process.env.PROJECT_NAME || configProject.projectName;
+        devHubAliasOk = (process.env.DEVHUB_ALIAS || configProject.devHubAlias) != null;
+        // If not found, prompt user project name and store it in user config file
+        if (projectName == null) {
+            const promptResponse = await prompts({
+                type: 'text',
+                name: 'value',
+                message: '[sfdx-hardis] Please input your project name without spaces or special characters (ex: MonClient)',
+                validate: (value: string) => !value.match(/^[0-9a-z]+$/) // check only alphanumeric
+            });
+            projectName = promptResponse.value;
+            await setConfig('project', {
+                projectName,
+                devHubAlias: `DevHub_${projectName}`
+            });
+            devHubAliasOk = true;
+        }
+    }
+
+    // Set DevHub username if not set
+    if (devHubAliasOk === false && options.Command && options.Command.supportsDevhubUsername === true) {
+        const configProject = await getConfig('project');
+        const devHubAlias = process.env.DEVHUB_ALIAS || configProject.devHubAlias;
+        if (devHubAlias == null) {
+            await setConfig('project', {
+                devHubAlias: `DevHub_${configProject.projectName}`
+            });
+        }
+    }
+};